Title: 범위 안의 수 2, Time: 140ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240305/범위 안의 수 2/number-in-range-2.java
@@ -18,7 +18,7 @@ public class Main {
             iterator++;
         }
         
-        average = sum / count;
+        average = (double)sum / count;
         System.out.print(sum + " ");
         System.out.printf("%.1f", average);
     }
